requirements.txt
torch
torchvision
numpy
pillow
scikit-image==0.19.3
scikit-learn==1.1.3
basicsr==1.4.2
facenet-pytorch==2.5.3
face-alignment==1.3.5
matplotlib==3.7.2